Synchronize initial state of client stores (#960)
Ref https://github.com/webstudio-is/webstudio-designer/issues/959

We usually rely on default which is the same everywhere. Though with
assets initial state comes from server so it should be synchronized.

The issue was not visible because we also load assets in runtime which
does not seem to happen on project switch.

Fixed by sending initial state to designer.